pyproject.toml
[project]
name = "resume-analyzer"
version = "2.0.0"
description = "Local AI Resume Analyzer using FLAN-T5"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "streamlit>=1.28.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
